feat: implement the same validation logic between log pages
