--- src/entities/ButtonWithRoute/ButtonWithRoute.css
@@ -1,17 +1,0 @@
-/* Error: Can't find stylesheet to import.
- *   ,
- * 1 | @import "src/constants/scss/variables";
- *   |         ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
- *   '
- *   ButtonWithRoute.scss 1:9  root stylesheet */
-
-body::before {
-  font-family: 'Source Code Pro', 'SF Mono', Monaco, Inconsolata, 'Fira Mono',
-    'Droid Sans Mono', monospace, monospace;
-  white-space: pre;
-  display: block;
-  padding: 1em;
-  margin-bottom: 1em;
-  border-bottom: 2px solid black;
-  content: 'Error: Can\'t find stylesheet to import.\a   \2577 \a 1 \2502  @import "src/constants/scss/variables";\d\a   \2502          ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^\a   \2575 \a   ButtonWithRoute.scss 1:9  root stylesheet';
-}
