--- src/entities/ButtonWithRoute/_ButtonWithRoute.scss
@@ -1,51 +1,0 @@
-@import 'src/constants/scss/variables';
-
-.button-signIn {
-  box-sizing: border-box;
-  padding: 10px 30px;
-  width: 140px;
-  height: 50px;
-  color: white;
-  background: $color-gradient-blue-orange;
-  font-size: 20px;
-  border: none;
-  line-height: 30px;
-  border-radius: 10px;
-  position: relative;
-  text-align: end;
-}
-
-.button-signIn a {
-  color: white;
-  text-decoration: none;
-}
-
-.button-signIn::before {
-  content: '';
-  position: absolute;
-  left: 10px;
-  top: 13px;
-  width: 24px;
-  height: 24px;
-  line-height: 50px;
-  background-image: url('../../assets/icons/icon-user.svg');
-  background-size: cover;
-  margin-right: 10px;
-}
-
-.button-signUp {
-  text-align: center;
-  width: 140px;
-  height: 50px;
-  line-height: 50px;
-  color: white;
-  background: linear-gradient(103.86deg, #00c6ff -22.01%, #0183e7 117.6%);
-  font-size: 20px;
-  border: none;
-  border-radius: 10px;
-}
-
-.button-signUp a {
-  color: white;
-  text-decoration: none;
-}
