--- src/entities/CartButton/CartButton.css
@@ -1,32 +1,0 @@
-.button-cart {
-  background: linear-gradient(to bottom left, #083fed, #fc8901);
-  width: 50px;
-  height: 50px;
-  border: none;
-  border-radius: 10px;
-  text-align: start;
-}
-
-.img-cart {
-  position: relative;
-  top: -7px;
-}
-
-.span-cart {
-  position: relative;
-  display: block;
-  text-align: center;
-  line-height: 20px;
-  font-size: 12px;
-  color: #fff;
-  width: 20px;
-  height: 20px;
-  left: 15px;
-  bottom: -3px;
-  background: linear-gradient(to bottom left, #083fed, #fc8901);
-  border: 1px solid white;
-  border-radius: 50%;
-  z-index: 10;
-}
-
-/*# sourceMappingURL=CartButton.css.map */
