--- src/entities/Categories/Categories.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import './Categories.scss';
-
-export function Categories() {
-  return <button className='button-categories'>Categories</button>;
-}
