--- src/entities/InputSubmit/InputSubmit.css
@@ -1,10 +1,0 @@
-.signIn_submit-button {
-  background-color: rgba(252, 137, 1, 0.8);
-  width: 270px;
-  height: 60px;
-  border: none;
-  color: white;
-  border-radius: 10px;
-}
-
-/*# sourceMappingURL=InputSubmit.css.map */
