--- src/entities/InputSubmit/InputSubmit.tsx
@@ -1,17 +1,0 @@
-import { FC } from 'react';
-import './InputSubmit.scss';
-
-interface InputSubmitProps {
-  className: string;
-  value: string;
-}
-
-const InputSubmit: FC<InputSubmitProps> = ({ className, value }) => {
-  return (
-    <>
-      <input className={className} type={'submit'} value={value} />
-    </>
-  );
-};
-
-export default InputSubmit;
