--- src/entities/SignUpButton/_SignUpButton.scss
@@ -1,14 +1,0 @@
-@import 'src/constants/scss/extentds';
-
-.button-signUp {
-  width: 140px;
-  height: 50px;
-  @extend %buttonSignUp;
-
-  a {
-    width: 100%;
-    height: 100%;
-    color: white;
-    text-decoration: none;
-  }
-}
