+++ src/features/formSubmitSignIn/FormSubmitSignIn.tsx
@@ -1,55 +1,67 @@
-import React, { useState } from 'react';
+import React from 'react';
+import { AnyAction, ThunkDispatch } from '@reduxjs/toolkit';
+import { useDispatch, useSelector } from 'react-redux';
 
-import InputSubmit from '../../entities/InputSubmit/InputSubmit';
-import InputWithValidation from '../InputWithValidation/InputWithValidation';
+import './_formSubmutSignIn.scss';
+
+import InputSubmit from '../../shared/components/InputSubmit/InputSubmit';
+import InputValidationSignIn from '../../entities/InputValidationSignIn/view/InputValidationSignIn';
 import {
   validateEmail,
   validatePassword,
-  validateSubmit,
-} from '../utils/validation/validationSignIn';
-import './_formSubmutSignIn.scss';
+} from '../../entities/InputValidationSignIn/usage/utils/validationSignIn';
 import { showPassword } from '../formCommon/showPassword';
+import { store } from '../../app/store/store';
+import { setSingInInputValidationError } from '../../app/store/signinAction/signinSlice';
+
+type RootState = ReturnType<typeof store.getState>;
 
 const FormSubmitSignIn = () => {
-  const [errorMessage, setErrorMessage] = useState('');
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
+  const dispatch: ThunkDispatch<RootState, unknown, AnyAction> = useDispatch();
+
+  const emailState = useSelector((state: RootState) => state.signin.email);
+
+  const passwordState = useSelector(
+    (state: RootState) => state.signin.password,
+  );
+
+  const checkInputEmail = () => {
+    const validationError: string = validateEmail(emailState.value);
+    const inputName = 'email';
+    dispatch(setSingInInputValidationError({ inputName, validationError }));
+  };
+
+  const checkInputPassword = () => {
+    const validationError: string = validatePassword(passwordState.value);
+    const inputName = 'password';
+    dispatch(setSingInInputValidationError({ inputName, validationError }));
+  };
 
   const handleFormSubmit = (event: React.FormEvent) => {
     event.preventDefault();
-    setErrorMessage(validateSubmit(email, password));
-  };
-
-  const handleInputEmail = (event: React.FormEvent<HTMLInputElement>) => {
-    const newValue: string = event.currentTarget.value;
-    setEmail(newValue);
-  };
-
-  const handleInputPassword = (event: React.FormEvent<HTMLInputElement>) => {
-    const newValue: string = event.currentTarget.value;
-    setPassword(newValue);
+    console.log(emailState);
+    // TODO: here not working
+    checkInputEmail();
+    checkInputPassword();
   };
 
   return (
     <form onSubmit={handleFormSubmit}>
-      <InputWithValidation
+      <InputValidationSignIn
         type={'text'}
         placeholder={'Email'}
-        value={email}
         valid={validateEmail}
-        onBlur={handleInputEmail}
+        inputName={'email'}
       />
-      <InputWithValidation
+      <InputValidationSignIn
         type={'password'}
         placeholder={'Password'}
         showPassword={showPassword}
-        value={password}
         valid={validatePassword}
-        onBlur={handleInputPassword}
+        inputName={'password'}
       />
       <div className={'container-submit'}>
         <InputSubmit className={'signIn_submit-button'} value={'SIGN IN'} />
-        <div className={'error-message'}>{errorMessage}</div>
       </div>
     </form>
   );
