+++ src/shared/components/Categories/Categories.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import './_Categories.scss';
+
+export function Categories() {
+  return <button className='button-categories'>Categories</button>;
+}
