+++ src/shared/components/InputCheckbox/InputCheckbox.tsx
