+++ src/shared/components/SignUpButton/SignUpButton.tsx
