+++ src/shared/components/SignUpButton/_SignUpButton.scss
@@ -1,0 +1,14 @@
+@import 'extentds';
+
+.button-signUp {
+  width: 140px;
+  height: 50px;
+  @extend %buttonSignUp;
+
+  a {
+    width: 100%;
+    height: 100%;
+    color: white;
+    text-decoration: none;
+  }
+}
