+++ src/widgets/WelcomeBackSection/WelcomeBackSection.tsx
@@ -1,4 +1,4 @@
-import ButtonWithRoute from '../../entities/ButtonWithRoute/ButtonWithRoute';
+import ButtonWithRoute from '../../shared/components/ButtonWithRoute/ButtonWithRoute';
 
 const WelcomeBackSection = () => {
   return (
